Assay Methods

Methods for Assay objects for generics defined in other packages

# S3 method for Assay
[(x, i, j, ...)

# S3 method for Assay
[[(x, i, ..., drop = FALSE)

# S3 method for Assay
dim(x)

# S3 method for Assay
dimnames(x)

# S3 method for Assay
merge(x = NULL, y = NULL, add.cell.ids = NULL, merge.data = TRUE, ...)

# S3 method for Assay
subset(x, cells = NULL, features = NULL, ...)

# S4 method for Assay
[[(x, i, j, ...) <- value

# S4 method for Assay
colMeans(x, na.rm = FALSE, dims = 1, ..., slot = "data")

# S4 method for Assay
colSums(x, na.rm = FALSE, dims = 1, ..., slot = "data")

# S4 method for Assay
rowMeans(x, na.rm = FALSE, dims = 1, ..., slot = "data")

# S4 method for Assay
rowSums(x, na.rm = FALSE, dims = 1, ..., slot = "data")

# S4 method for Assay
show(object)

Arguments

x, object
An Assay object

i, features
For [[: metadata names; for all other methods, feature names or indices

j, cells
Cell names or indices

...
Arguments passed to other methods

drop
See drop

y
A vector or list of one or more objects to merge

add.cell.ids
A character vector of length(x = c(x, y)); appends the corresponding values to the start of each objects' cell names

merge.data
Merge the data slots instead of just merging the counts (which requires renormalization); this is recommended if the same normalization approach was applied to all objects

value
Additional metadata to add

na.rm
logical. Should missing values (including NaN) be omitted from the calculations?

dims
completely ignored by the Matrix methods.

slot
Specific information to pull (i.e. counts, data, scale.data, ...)

Value

[: The data slot for features i and cells j

[[: The feature-level metadata for i

dim: The number of features (nrow) and cells (ncol)

dimnames: Feature (row) and cell (column) names

merge: Merged object

subset: A subsetted Assay

[[<-: x with metadata value added as i

colMeans: The column (cell-wise) means of slot

colSums: The column (cell-wise) sums of slot

rowMeans: The row (feature-wise) means of slot

rowSums: The row (feature-wise) sums of slot

show: Prints summary to stdout and invisibly returns NULL

Functions

[.Assay: Get expression data from an Assay

[[.Assay: Get feature-level metadata

dim.Assay: Number of cells and features for an Assay

dimnames.Assay: Cell- and feature-names for an Assay

merge.Assay: Merge Assay objects

subset.Assay: Subset an Assay

[[<-,Assay-method: Add feature-level metadata

colMeans,Assay-method: Calculate colMeans on an Assay

colSums,Assay-method: Calculate colSums on an Assay

rowMeans,Assay-method: Calculate rowMeans on an Assay

rowSums,Assay-method: Calculate rowSums on an Assay

show,Assay-method: Overview of an Assay object