The Transposed v5 Assay Object

The Transposed v5 Assay Object

Slots

layers

A named list containing expression matrices; each matrix should be a two-dimensional object containing some subset of cells and features defined in the cells and features slots. Cell and feature membership is recorded in the cells and features slots, respectively

cells

A logical mapping of cell names and layer membership; this map contains all the possible cells that this assay can contain. New layers must have some subset of cells present in this map

features

A logical mapping of feature names and layer membership; this map contains all the possible features that this assay can contain. New layers must have some subset of features present in this map

default

A one-length integer with the end index of the default layer; the default layer be all layers up to and including the layer at index default

assay.orig

Original assay that this assay is based off of; used to track assay provenance

meta.data

A data frame with feature-level meta data; should have the same number of rows as features

misc

A named list of unstructured miscellaneous data

key

A one-length character vector with the object's key; keys must be one or more alphanumeric characters followed by an underscore “_” (regex pattern “^[a-zA-Z][a-zA-Z0-9]*_$ ”)

Lifecycle

Warning: functionality described here is experimental and prone to change without notice