Get and Set Assay Data

General accessor and setter functions for Assay objects. GetAssayData can be used to pull information from any of the expression matrices (eg. “counts”, “data”, or “scale.data”). SetAssayData can be used to replace one of these expression matrices

# S3 method for StdAssay
GetAssayData(object, layer = NULL, slot = deprecated(), ...)

# S3 method for StdAssay
SetAssayData(object, layer, new.data, slot = deprecated(), ...)

Arguments

object

An object

layer

Name of layer to get or set

slot

Specific assay data to get or set

...

Arguments passed to other methods

new.data

New assay data to add

Value

GetAssayData: returns the specified assay data

SetAssayData: object with the assay data set

Lifecycle

GetAssayData and SetAssayData have been superseded. To fetch expression matrices, use LayerData; to set expression data, use LayerData<-