Get cell names grouped by identity class

Get cell names grouped by identity class

CellsByIdentities(object, idents = NULL, cells = NULL)

Arguments

object
A Seurat object

idents
A vector of identity class levels to limit resulting list to; defaults to all identity class levels

cells
A vector of cells to grouping to

Value

A named list where names are identity classes and values are vectors of cells belonging to that class

Examples

CellsByIdentities(object = pbmc_small)

#> $`0`
#>  [1] "ATGCCAGAACGACT" "CATGGCCTGTGCAT" "GAACCTGATGAACC" "TGACTGGATTCTCA"
#>  [5] "AGTCAGACTGCACA" "TCTGATACACGTGT" "TGGTATCTAAACAG" "GCAGCTCTGTTTCT"
#>  [9] "GATATAACACGCAT" "AATGTTGACAGTCA" "CTAAACCTCTGACA" "CTAACGGAACCGAT"
#> [13] "AGATATACCCGTAA" "TACTCTGAATCGAC" "GCGCATCTTGCTCC" "GTTGACGATATCGG"
#> [17] "ACAGGTACTGGTGT" "GGCATATGCTTATC" "CATTACACCAACTG" "TAGGGACTGAACTC"
#> [21] "TACAATGATGCTAG" "CTTCATGACCGAAT" "AAGCAAGAGCTTAG" "CGGCACGAACTCAG"
#> [25] "GGTGGAGATTACTC" "GGCCGATGTACTCT" "CATGCGCTAGTCAC" "ATTCAGCTCATTGG"
#> [29] "GGCATATGGGGAGT" "ATCATCTGACACCA" "GTCATACTTCGCCT" "TTACGTACGTTCAG"
#> [33] "GAGTTGTGGTAGCT" "GACGCTCTCTCTCG" "AGTCTTACTTCGGA" "GGAACACTTCAGAC"
#> 
#> $`1`
#>  [1] "TTACCATGAATCGC" "ATAGGAGAAACAGA" "GCGCACGACTTTAC" "ACTCGCACGAAAGT"
#>  [5] "ATTACCTGCCTTAT" "CCCAACTGCAATCG" "AAATTCGAATCACG" "CCATCCGATTCGCC"
#>  [9] "TCCACTCTGAGCTT" "CATCAGGATGCACA" "CGTAGCCTGTATGC" "TGAGCTGAATGCTG"
#> [13] "ATAAGTTGGTACGT" "AAGCGACTTTGACG" "ACCAGTGAATACCG" "ATTGCACTTGCTTT"
#> [17] "CTAGGTGATGGTTG" "GCACTAGACCTTTA" "ATACCACTCTAAGC" "TTTAGCTGTACTCT"
#> [21] "ACGTGATGCCATGA" "ATTGTAGATTCCCG" "GATAGAGATCACGA" "AATGCGTGGACGGA"
#> [25] "CTTGATTGATCTTC"
#> 
#> $`2`
#>  [1] "AGGTCATGAGTGTC" "AGAGATGATCTCGC" "GGGTAACTCTAGTG" "CATGAGACACGGGA"
#>  [5] "TACGCCACTCCGAA" "CTAAACCTGTGCAT" "GTAAGCACTCATTC" "TTGGTACTGAATCC"
#>  [9] "CATCATACGGAGCA" "TACATCACGCTAAC" "GATAGAGAAGGGTG" "GCTCCATGAGAAGT"
#> [13] "CTGCCAACAGGAGC" "TTGCATTGAGCTAC" "CCTATAACGAGACG" "TTGAGGACTACGCA"
#> [17] "CATATAGACTAAGC" "GACATTCTCCACCT" "GCGTAAACACGGTT"
#>