Create an Assay object

Create an Assay object from a feature (e.g. gene) expression matrix. The expected format of the input matrix is features x cells.

CreateAssayObject(counts, data, min.cells = 0, min.features = 0)

Arguments

counts
Unnormalized data such as raw counts or TPMs

data
Prenormalized data; if provided, do not pass counts

min.cells
Include features detected in at least this many cells. Will subset the counts matrix as well. To reintroduce excluded features, create a new object with a lower cutoff.

min.features
Include cells where at least this many features are detected.

Details

Non-unique cell or feature names are not allowed. Please make unique before calling this function.

Examples

pbmc_raw <- read.table(
  file = system.file('extdata', 'pbmc_raw.txt', package = 'Seurat'),
  as.is = TRUE
)

#> Warning: file("") only supports open = "w+" and open = "w+b": using the former

#> Error in read.table(file = system.file("extdata", "pbmc_raw.txt", package = "Seurat"),     as.is = TRUE): no lines available in input

pbmc_rna <- CreateAssayObject(counts = pbmc_raw)

#> Error in rownames(x = counts): object 'pbmc_raw' not found

pbmc_rna

#> Error in eval(expr, envir, enclos): object 'pbmc_rna' not found