Create Spatial Coordinates

Create Spatial Coordinates

CreateFOV(coords, ...)

# S3 method for Centroids
CreateFOV(
  coords,
  molecules = NULL,
  assay = "Spatial",
  key = NULL,
  name = NULL,
  ...
)

# S3 method for data.frame
CreateFOV(
  coords,
  type = c("segmentation", "centroids"),
  nsides = Inf,
  radius = NULL,
  theta = 0L,
  molecules = NULL,
  assay = "Spatial",
  key = NULL,
  name = NULL,
  ...
)

# S3 method for list
CreateFOV(coords, molecules = NULL, assay = "Spatial", key = NULL, ...)

# S3 method for Segmentation
CreateFOV(
  coords,
  molecules = NULL,
  assay = "Spatial",
  key = NULL,
  name = NULL,
  ...
)

Arguments

coords

Spatial coordinates

...

Arguments passed to other methods

molecules

A data.frame with spatially-resolved molecule information or a Molecules object

assay

Name of associated assay

key

Key for these spatial coordinates

name

When coords is a data.frame, Centroids, or Segmentation, name to store coordinates as

type

When providing a data.frame, specify if the coordinates represent a cell segmentation or voxel centroids

nsides

The number of sides to represent cells/spots; pass Inf to plot as circles

radius

Radius of shapes when plotting

theta

Angle to adjust shapes when plotting

Value

A FOV object

See also