Create a Seurat object

Create a Seurat object from raw data

CreateSeuratObject(
  counts,
  project = "CreateSeuratObject",
  assay = "RNA",
  names.field = 1,
  names.delim = "_",
  meta.data = NULL,
  ...
)

# S3 method for default
CreateSeuratObject(
  counts,
  project = "SeuratProject",
  assay = "RNA",
  names.field = 1,
  names.delim = "_",
  meta.data = NULL,
  min.cells = 0,
  min.features = 0,
  ...
)

# S3 method for Assay
CreateSeuratObject(
  counts,
  project = "SeuratProject",
  assay = "RNA",
  names.field = 1,
  names.delim = "_",
  meta.data = NULL,
  ...
)

Arguments

counts
Either a matrix-like object with unnormalized data with cells as columns and features as rows or an Assay-derived object

project
Project name for the Seurat object

assay
Name of the initial assay

names.field
For the initial identity class for each cell, choose this field from the cell's name. E.g. If your cells are named as BARCODE_CLUSTER_CELLTYPE in the input matrix, set names.field to 3 to set the initial identities to CELLTYPE.

names.delim
For the initial identity class for each cell, choose this delimiter from the cell's column name. E.g. If your cells are named as BARCODE-CLUSTER-CELLTYPE, set this to “-” to separate the cell name into its component parts for picking the relevant field.

meta.data
Additional cell-level metadata to add to the Seurat object. Should be a data.frame where the rows are cell names and the columns are additional metadata fields.

...
Arguments passed to other methods

Value

A Seurat object

Note

In previous versions (<3.0), this function also accepted a parameter to set the expression threshold for a ‘detected’ feature (gene). This functionality has been removed to simplify the initialization process/assumptions. If you would still like to impose this threshold for your particular dataset, simply filter the input expression matrix before calling this function.

Examples

pbmc_raw <- read.table(
  file = system.file('extdata', 'pbmc_raw.txt', package = 'Seurat'),
  as.is = TRUE
)

#> Warning: file("") only supports open = "w+" and open = "w+b": using the former

#> Error in read.table(file = system.file("extdata", "pbmc_raw.txt", package = "Seurat"),     as.is = TRUE): no lines available in input

pbmc_small <- CreateSeuratObject(counts = pbmc_raw)

#> Error in CreateSeuratObject(counts = pbmc_raw): object 'pbmc_raw' not found

pbmc_small

#> Error in eval(expr, envir, enclos): object 'pbmc_small' not found