DimReduc Methods

Methods for DimReduc objects for generics defined in other packages

# S3 method for DimReduc
[(x, i, j, drop = FALSE, ...)

# S3 method for DimReduc
[[(x, i, j, drop = FALSE, ...)

# S3 method for DimReduc
dim(x)

# S3 method for DimReduc
dimnames(x)

# S3 method for DimReduc
length(x)

# S3 method for DimReduc
merge(x = NULL, y = NULL, add.cell.ids = NULL, ...)

# S3 method for DimReduc
names(x)

# S3 method for DimReduc
print(x, dims = 1:5, nfeatures = 20, projected = FALSE, ...)

# S3 method for DimReduc
subset(x, cells = NULL, features = NULL, ...)

# S4 method for DimReduc
show(object)

Arguments

x, object
A DimReduc object

i
For [: feature names or indices; for [[: cell names or indices

j
Dimensions to pull for

drop
See drop

...
Arguments passed to other methods

y
A vector or list of one or more objects to merge

add.cell.ids
A character vector of length(x = c(x, y)); appends the corresponding values to the start of each objects' cell names

dims
Number of dimensions to display

nfeatures
Number of genes to display

projected
Use projected slot

cells, features
Cells and features to keep during the subset

Value

[: Feature loadings for features i and dimensions j

[[: Cell embeddings for cells i and dimensions j

dim: The number of cells (nrow) and dimensions (ncol)

dimnames: The cell (row) and dimension (column) names

length: The number of dimensions

names: The names for the dimensions (eg. “PC_1”)

print: Displays set of features defining the components and invisibly returns x

subset: x for cells cells and features features

show: Prints summary to stdout and invisibly returns NULL

Functions

[.DimReduc: Pull feature loadings

[[.DimReduc: Pull cell embeddings

dim.DimReduc: The number of cells and dimensions for a DimReduc

dimnames.DimReduc: The cell and dimension names for a DimReduc object

length.DimReduc: The number of dimensions for a DimReduc object

merge.DimReduc: Merge two or more DimReduc objects together

names.DimReduc: The dimension names for a DimReduc object

print.DimReduc: Prints a set of features that most strongly define a set of components; note: requires feature loadings to be present in order to work

subset.DimReduc: Subset a DimReduc object

show,DimReduc-method: Show basic summary of a DimReduc object

See also