Dimensional Reduction Validity

Validation of DimReduc objects is handled by validObject

Cell Embeddings Validation

The cell embeddings matrix must be a numeric matrix of dimensions \(n_{cells}\) by \(d_{dimensions}\); row names must be the cell names and column names must be the dimension identifier. The dimension identifier must be “key_dimension” (eg. “PC_1”). Dimension identifiers must be in order and cannot be skipped

Feature and Projected Feature Loadings Validation

blah

Standard Deviations Validation

blah

Key Validation

Keys must be a one-length character vector; a key must be composed of one of the following:

An empty string (eg. “''”) where nzchar() == 0

An string composed of one or more alphanumeric values (both lower- and upper-case) that ends with an underscore (“_”); the first character must be a letter

Keys that are not empty strings are validated with the regex “^[a-zA-Z][a-zA-Z0-9]*_$ ”

See also