A Mixin for Keyed objects

A mixin (virtual class) for enabling keyed objects; provides consistent behavior for getting, setting, and validating keys

# S3 method for character
Key(object, quiet = FALSE, ...)

# S3 method for KeyMixin
Key(object, ...)

# S3 method for KeyMixin
Key(object, ...) <- value

Arguments

object

An object

quiet

Suppress warnings when updating characters to keys

...

Ignored

value

A key to set

Value

Key.character: object but as a syntactically-valid key

Key.KeyMixin: The key from object

Key<-: object with the key set to value

Details

Key.character: Update a character to a key

Key.KeyMixin: Get the key of a keyed object

Key<-: Set the key of a keyed object

Slots

key

A one-length character vector with the object's key; keys must be one or more alphanumeric characters followed by an underscore “_” (regex pattern “^[[:alnum:]]+_$”)