Query and Manipulate Assay Layers

Query and Manipulate Assay Layers

LayerData(object, layer, ...)

LayerData(object, layer, ...) <- value

Layers(object, ...)

# S3 method for Assay
LayerData(
  object,
  layer = NULL,
  cells = NULL,
  features = NULL,
  slot = deprecated(),
  ...
)

# S3 method for Assay
LayerData(object, layer, ...) <- value

# S3 method for Assay
Layers(object, search = NA, ...)

# S3 method for Assay5
LayerData(
  object,
  layer = NULL,
  cells = NULL,
  features = NULL,
  fast = FALSE,
  slot = deprecated(),
  ...
)

# S3 method for Assay5
LayerData(object, layer, features = NULL, cells = NULL, ...) <- value

# S3 method for Assay5
Layers(object, search = NA, ...)

# S3 method for Seurat
LayerData(object, layer = NULL, assay = NULL, slot = deprecated(), ...)

# S3 method for Seurat
LayerData(object, layer, assay = NULL, ...) <- value

# S3 method for Seurat
Layers(object, search = NA, assay = NULL, ...)

Arguments

object

An object

layer

Name of layer to fetch or set

...

Arguments passed to other methods

value

New two-dimensional data to be added as a layer

features, cells

Vectors of features/cells to include

slot

search

A pattern to search layer names for; pass one of:

“NA” to pull all layers

“NULL” to pull the default layer(s)

a regular expression that matches layer names

fast

Determine how to return the layer data; choose from:

FALSE

Apply any transpositions and attempt to add feature/cell names (if supported) back to the layer data

NA

Attempt to add feature/cell names back to the layer data, skip any transpositions

TRUE

Do not apply any transpositions or add feature/cell names to the layer data

assay

Name of assay to fetch layer data from or assign layer data to

Value

LayerData: the layer data for layer from object

Layer<-: object with value added as a layer named layer

Layers: the names of the layers present in object