The Seurat Class

The Seurat object is a representation of single-cell expression data for R; each Seurat object revolves around a set of cells and consists of one or more Assay objects, or individual representations of expression data (eg. RNA-seq, ATAC-seq, etc). These assays can be reduced from their high-dimensional state to a lower-dimension state and stored as DimReduc objects. Seurat objects also store additional metadata, both at the cell and feature level (contained within individual assays). The object was designed to be as self-contained as possible, and easily extendable to new methods.

Slots

assays

A list of assays for this project

meta.data

Contains meta-information about each cell, starting with number of features detected (nFeature) and the original identity class (orig.ident); more information is added using AddMetaData

active.assay

Name of the active, or default, assay; settable using DefaultAssay

active.ident

The active cluster identity for this Seurat object; settable using Idents

graphs

A list of Graph objects

neighbors

...

reductions

A list of dimensional reduction objects for this object

images

A list of spatial image objects

project.name

Name of the project

misc

A list of miscellaneous information

version

Version of Seurat this object was built under

commands

A list of logged commands run on this Seurat object

tools

A list of miscellaneous data generated by other tools, should be filled by developers only using Tool<-