Seurat Methods

Methods for Seurat objects for generics defined in other packages

# S3 method for Seurat
.DollarNames(x, pattern = "")

# S3 method for Seurat
$(x, i, ...)

# S3 method for Seurat
$(x, i, ...) <- value

# S3 method for Seurat
[(x, i, j, ...)

# S3 method for Seurat
[[(x, i, ..., drop = FALSE)

# S3 method for Seurat
dim(x)

# S3 method for Seurat
dimnames(x)

# S3 method for Seurat
names(x)

# S4 method for Seurat
[[(x, i, j, ...) <- value

# S4 method for Seurat
show(object)

Arguments

x, object
A Seurat object

pattern
A regular expression. Only matching names are returned.

i, features
Depends on the method

[, subset

Feature names or indices

$, $<-

Name of a single metadata column

[[, [[<-

Name of one or more metadata columns or an associated object; associated objects include Assay, DimReduc, Graph, SeuratCommand, or SpatialImage objects

@param j,cells Cell names or indices @param ... Arguments passed to other methods

value
Additional metadata or associated objects to add; note: can pass NULL to remove metadata or an associated object

drop
See drop

Value

$: metadata column i for object x; note: unlike [[, $ drops the shape of the metadata to return a vector instead of a dataframe

$<-: object x with metadata value saved as i

[: object x with features i and cells j

[[: If i is missing, the metadata dataframe; if i is a vector of metadata names, a dataframe with the requested metadata, otherwise, the requested associated object

dim: The number of features (nrow) and cells (ncol) for the default assay; note: while the number of features changes depending on the active assay, the number of cells remains the same across all assays

dimnames: The feature (row) and cell (column) names; note: while the features change depending on the active assay, the cell names remain the same across all assays

names: The names of all Assay, DimReduc, Graph, and SpatialImage objects in the Seurat object

[[<-: x with the metadata or associated objects added as i; if value is NULL, removes metadata or associated object i from object x

show: Prints summary to stdout and invisibly returns NULL

Functions

.DollarNames.Seurat: Autocompletion for $ access on a Seurat object

$.Seurat: Metadata access for Seurat objects

$<-.Seurat: Metadata setter for Seurat objects

[.Seurat: Simple subsetter for Seurat objects

[[.Seurat: Metadata and associated object accessor

dim.Seurat: Number of cells and features for the active assay

dimnames.Seurat: The cell and feature names for the active assay

names.Seurat: Common associated objects

[[<-,Seurat-method: Add cell-level metadata or associated objects

show,Seurat-method: Overview of a Seurat object

Examples

pbmc_small[VariableFeatures(object = pbmc_small), ]

#> Error in eval(expr, envir, enclos): object 'pbmc_small' not found

pbmc_small[, 1:10]

#> Error in eval(expr, envir, enclos): object 'pbmc_small' not found