SeuratObject: Data Structures for Single Cell Data

Defines S4 classes for single-cell genomic data and associated information, such as dimensionality reduction embeddings, nearest-neighbor graphs, and spatially-resolved coordinates. Provides data access methods and R-native hooks to ensure the Seurat object is familiar to other R users. See Satija R, Farrell J, Gennert D, et al (2015) <doi:10.1038/nbt.3192>, Macosko E, Basu A, Satija R, et al (2015) <doi:10.1016/j.cell.2015.05.002>, and Stuart T, Butler A, et al (2019) <doi:10.1016/j.cell.2019.05.031> for more details.

See also

Author

Maintainer: Paul Hoffman nygcSatijalab@nygenome.org (ORCID)

Authors:

Rahul Satija rsatija@nygenome.org (ORCID)

Andrew Butler abutler@nygenome.org (ORCID)

Tim Stuart tstuart@nygenome.org (ORCID)

Other contributors:

Jeff Farrell jfarrell@g.harvard.edu [contributor]

Shiwei Zheng szheng@nygenome.org (ORCID) [contributor]

Christoph Hafemeister chafemeister@nygenome.org (ORCID) [contributor]

Patrick Roelli proelli@nygenome.org [contributor]

Yuhan Hao yhao@nygenome.org (ORCID) [contributor]