The SpatialImage class

The SpatialImage class is a virtual class representing spatial information for Seurat. All spatial image information must inherit from this class for use with Seurat objects

# S3 method for SpatialImage
Cells(x)

# S3 method for SpatialImage
GetImage(object, mode = c("grob", "raster", "plotly", "raw"), ...)

# S3 method for SpatialImage
GetTissueCoordinates(object, ...)

# S3 method for SpatialImage
Radius(object)

# S3 method for SpatialImage
RenameCells(object, new.names = NULL, ...)

# S3 method for SpatialImage
[(x, i, ...)

# S3 method for SpatialImage
dim(x)

# S3 method for SpatialImage
subset(x, cells, ...)

Arguments

x, object
An object inheriting from SpatialImage

mode
How to return the image; should accept one of 'grob', 'raster', 'plotly', or 'raw'

...
Arguments passed to other methods

new.names
vector of new cell names

i, cells
A vector of cells to keep

Slots

assay

Name of assay to associate image data with; will give this image priority for visualization when the assay is set as the active/default assay in a Seurat object

key

Key for the image

Provided methods

These methods are defined on the SpatialImage object and should not be overwritten without careful thought

DefaultAssay and DefaultAssay<-

Key and Key<-

IsGlobal

Radius; this method can be overridden to provide a spot radius for image objects

Required methods

All subclasses of the SpatialImage class must define the following methods; simply relying on the SpatialImage method will result in errors. For required parameters and their values, see the Usage and Arguments sections

Cells

Return the cell/spot barcodes associated with each position

dim

Return the dimensions of the image for plotting in (Y, X) format

GetImage

Return image data; by default, must return a grob object

GetTissueCoordinates

Return tissue coordinates; by default, must return a two-column data.frame with x-coordinates in the first column and y-coordiantes in the second

Radius

Return the spot radius; returns NULL by default for use with non-spot image technologies

RenameCells

Rename the cell/spot barcodes for this image

subset and [

Subset the image data by cells/spots; [ should only take i for subsetting by cells/spots

These methods are used throughout Seurat, so defining them and setting the proper defaults will allow subclasses of SpatialImage to work seamlessly