Find features with highest scores for a given dimensional reduction technique

Return a list of features with the strongest contribution to a set of components

TopFeatures(
  object,
  dim = 1,
  nfeatures = 20,
  projected = FALSE,
  balanced = FALSE,
  ...
)

Arguments

object
DimReduc object

dim
Dimension to use

nfeatures
Number of features to return

projected
Use the projected feature loadings

balanced
Return an equal number of features with both + and - scores.

...
Extra parameters passed to Loadings

Value

Returns a vector of features

Examples

pbmc_small

#> Error in eval(expr, envir, enclos): object 'pbmc_small' not found

TopFeatures(object = pbmc_small[["pca"]], dim = 1)

#> Error in Loadings(object = object, projected = projected, ...): object 'pbmc_small' not found

# After projection:
TopFeatures(object = pbmc_small[["pca"]], dim = 1,  projected = TRUE)

#> Error in Loadings(object = object, projected = projected, ...): object 'pbmc_small' not found