Highly Variable Features

Get and set variable feature information for an Assay object. HVFInfo and VariableFeatures utilize generally variable features, while SVFInfo and SpatiallyVariableFeatures are restricted to spatially variable features

HVFInfo(object, selection.method, status = FALSE, ...)

VariableFeatures(object, selection.method = NULL, ...)

VariableFeatures(object, ...) <- value

SVFInfo(object, selection.method, status, ...)

SpatiallyVariableFeatures(object, selection.method, ...)

# S3 method for Seurat
HVFInfo(object, selection.method = NULL, status = FALSE, assay = NULL, ...)

# S3 method for Seurat
VariableFeatures(object, selection.method = NULL, assay = NULL, ...)

# S3 method for Seurat
VariableFeatures(object, assay = NULL, ...) <- value

# S3 method for Seurat
SVFInfo(
  object,
  selection.method = c("markvariogram", "moransi"),
  status = FALSE,
  assay = NULL,
  ...
)

# S3 method for Seurat
SpatiallyVariableFeatures(
  object,
  selection.method = "markvariogram",
  assay = NULL,
  decreasing = TRUE,
  ...
)

# S3 method for Assay
HVFInfo(object, selection.method, status = FALSE, ...)

# S3 method for Assay
SpatiallyVariableFeatures(
  object,
  selection.method = "markvariogram",
  decreasing = TRUE,
  ...
)

# S3 method for Assay
SVFInfo(
  object,
  selection.method = c("markvariogram", "moransi"),
  status = FALSE,
  ...
)

# S3 method for Assay
VariableFeatures(object, selection.method = NULL, ...)

# S3 method for Assay
VariableFeatures(object, ...) <- value

Arguments

object

An object

selection.method

Which method to pull. For HVFInfo and VariableFeatures, choose one from one of the following:

“vst”

“sctransform” or “sct”

“mean.var.plot”, “dispersion”, “mvp”, or “disp”

For SVFInfo and SpatiallyVariableFeatures, choose from:

“markvariogram”

“moransi”

status

Add variable status to the resulting data frame

...

Arguments passed to other methods

value

A character vector of variable features

assay

Name of assay to pull highly variable feature information for

decreasing

Return features in decreasing order (most spatially variable first).

Value

HVFInfo: A data frame with feature means, dispersion, and scaled dispersion

VariableFeatures: a vector of the variable features

SVFInfo: a data frame with the spatially variable features

SpatiallyVariableFeatures: a character vector of the spatially variable features

Examples

# Get the HVF info from a specific Assay in a Seurat object
HVFInfo(object = pbmc_small, assay = "RNA")[1:5, ]
#>            mean    variance variance.standardized
#> MS4A1    0.3875   1.0251582             0.8983463
#> CD79B    0.6000   1.2810127             0.4731134
#> CD79A    0.7000   4.3645570             1.0862810
#> HLA-DRA 13.4250 725.4626582             0.9730883
#> TCL1A    0.3000   0.8708861             1.3558627

# Get the HVF info directly from an Assay object
HVFInfo(pbmc_small[["RNA"]], selection.method = 'vst')[1:5, ]
#>            mean    variance variance.standardized
#> MS4A1    0.3875   1.0251582             0.8983463
#> CD79B    0.6000   1.2810127             0.4731134
#> CD79A    0.7000   4.3645570             1.0862810
#> HLA-DRA 13.4250 725.4626582             0.9730883
#> TCL1A    0.3000   0.8708861             1.3558627