Identify cells matching certain criteria

Returns a list of cells that match a particular set of criteria such as identity class, high/low values for particular PCs, ect..

WhichCells(object, ...)

# S3 method for Assay
WhichCells(object, cells = NULL, expression, invert = FALSE, ...)

# S3 method for Seurat
WhichCells(
  object,
  cells = NULL,
  idents = NULL,
  expression,
  slot = "data",
  invert = FALSE,
  downsample = Inf,
  seed = 1,
  ...
)

Arguments

object
An object

...
Arguments passed to other methods

cells
Subset of cell names

expression
A predicate expression for feature/variable expression, can evaluate anything that can be pulled by FetchData; please note, you may need to wrap feature names in backticks (``) if dashes between numbers are present in the feature name

invert
Invert the selection of cells

idents
A vector of identity classes to keep

slot
Slot to pull feature data for

downsample
Maximum number of cells per identity class, default is Inf; downsampling will happen after all other operations, including inverting the cell selection

seed
Random seed for downsampling. If NULL, does not set a seed

Value

A vector of cell names

See also

Examples

WhichCells(object = pbmc_small, idents = 2)

#> Error in WhichCells(object = pbmc_small, idents = 2): object 'pbmc_small' not found

WhichCells(object = pbmc_small, expression = MS4A1 > 3)

#> Error in WhichCells(object = pbmc_small, expression = MS4A1 > 3): object 'pbmc_small' not found

levels(x = pbmc_small)

#> Error in levels(x = pbmc_small): object 'pbmc_small' not found

WhichCells(object = pbmc_small, idents = c(1, 2), invert = TRUE)

#> Error in WhichCells(object = pbmc_small, idents = c(1, 2), invert = TRUE): object 'pbmc_small' not found