Aggregate Molecules into an Expression Matrix

Aggregate Molecules into an Expression Matrix

# S3 method for FOV
aggregate(x, by = NULL, set = NULL, drop = TRUE, ...)

# S3 method for Molecules
aggregate(x, by, drop = TRUE, ...)

Arguments

x

An object with spatially-resolved molecule information

by

Name of a Segmentation within object or a Segmentation object

set

Name of molecule set to aggregate

drop

Drop molecules not present in a segmentation; if FALSE, adds a column called “boundless” consisting of molecule counts not in a segmentation

...

Arguments passed to other methods

Value

An expression matrix

Progress Updates with progressr

This function uses progressr to render status updates and progress bars. To enable progress updates, wrap the function call in with_progress or run handlers(global = TRUE) before running this function. For more details about progressr, please read vignette("progressr-intro")

Parallelization with future

This function uses future to enable parallelization. Parallelization strategies can be set using plan. Common plans include “sequential” for non-parallelized processing or “multisession” for parallel evaluation using multiple R sessions; for other plans, see the “Implemented evaluation strategies” section of ?future::plan. For a more thorough introduction to future, see vignette("future-1-overview")