Update a Class's Package

Swap packages for an object's class definition. As classes move between packages, these functions rescope the namespace of the S4 class. This allows objects to depend only on the new package for class definitions rather than both the new and old packages

SwapClassPkg(x, from = NULL, to = NULL)

UpdateClassPkg(object, from = NULL, to = NULL)

Arguments

x
A list with an S4 class definition attribute

from
A vector of one or more packages to limit conversion from

to
A character naming the package to search for new class definitions; defaults to the package of the function calling this function

object
An S4 object

Value

SwapClassPkg: x with an updated S4 class definition attribute

UpdateClassPkg: object with the updated class definition

S4 Class Definition Attributes

S4 classes are scoped to the package and class name. In order to properly track which class a list is generated from in order to build a new one, these function use an attribute to denote the class name and package of origin. This attribute is stored as “classDef” and takes the form of “package:class”.

See also