S4/List Conversion

Convert S4 objects to lists and vice versa. Useful for declassing an S4 object while keeping track of it's class using attributes (see section S4 Class Definition Attributes below for more details). Both ListToS4 and S4ToList are recursive functions, affecting all lists/S4 objects contained as sub-lists/sub-objects

S4ToList(object)

IsS4List(x)

ListToS4(x)

# S3 method for default
S4ToList(object)

# S3 method for list
S4ToList(object)

Arguments

object

An S4 object

x

A list with an S4 class definition (“classDef”) attribute

Value

S4ToList: A list with an S4 class definition attribute

IsS4List: TRUE if x is a list with an S4 class definition attribute

ListToS4: An S4 object as defined by the S4 class definition attribute

S4 Class Definition Attributes

S4 classes are scoped to the package and class name. In order to properly track which class a list is generated from in order to build a new one, these function use an attribute to denote the class name and package of origin. This attribute is stored as 1-length character vector named “classDef” and takes the form of “package:class”

See also

Examples

# Turn an S4 object into a list
pbmc.list <- S4ToList(pbmc_small)
class(pbmc.list)
#> [1] "list"
attributes(pbmc.list)
#> $names
#>  [1] "assays"       "meta.data"    "active.assay" "active.ident" "graphs"      
#>  [6] "neighbors"    "reductions"   "images"       "project.name" "misc"        
#> [11] "version"      "commands"     "tools"       
#> 
#> $classDef
#> [1] "SeuratObject:Seurat"
#> 

str(pbmc.list$reductions)
#> List of 2
#>  $ pca :List of 9
#>   ..$ cell.embeddings           : num [1:80, 1:19] -0.774 -0.026 -0.457 -0.812 -0.774 ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : chr [1:80] "ATGCCAGAACGACT" "CATGGCCTGTGCAT" "GAACCTGATGAACC" "TGACTGGATTCTCA" ...
#>   .. .. ..$ : chr [1:19] "PC_1" "PC_2" "PC_3" "PC_4" ...
#>   ..$ feature.loadings          : num [1:20, 1:19] 0.3383 -0.035 0.1199 -0.0469 -0.0389 ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : chr [1:20] "PPBP" "IGLL5" "VDAC3" "CD1C" ...
#>   .. .. ..$ : chr [1:19] "PC_1" "PC_2" "PC_3" "PC_4" ...
#>   ..$ feature.loadings.projected: num [1:20, 1:19] -21.5 -38.8 -45 -75.7 -63.7 ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : chr [1:20] "IGLL5" "S100A8" "S100A9" "HLA-DPB1" ...
#>   .. .. ..$ : chr [1:19] "PC_1" "PC_2" "PC_3" "PC_4" ...
#>   ..$ assay.used                : chr "RNA"
#>   ..$ global                    : logi FALSE
#>   ..$ stdev                     : num [1:19] 2.79 1.61 1.32 1.12 1.03 ...
#>   ..$ jackstraw                 :List of 4
#>   .. ..$ empirical.p.values     : num [1:20, 1:10] 0 0.467 0 0.333 0.4 ...
#>   .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. ..$ : chr [1:20] "PPBP" "IGLL5" "VDAC3" "CD1C" ...
#>   .. .. .. ..$ : chr [1:10] "PC1" "PC2" "PC3" "PC4" ...
#>   .. ..$ fake.reduction.scores  : num [1:30, 1:10] 0.0408 -0.02849 -0.00419 -0.02824 0.01626 ...
#>   .. ..$ empirical.p.values.full: logi [1, 1] NA
#>   .. ..$ overall.p.values       : num [1:5, 1:2] 1 2 3 4 5 ...
#>   .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. ..$ : NULL
#>   .. .. .. ..$ : chr [1:2] "PC" "Score"
#>   .. ..- attr(*, "classDef")= chr "SeuratObject:JackStrawData"
#>   ..$ misc                      : list()
#>   ..$ key                       : chr "PC_"
#>   ..- attr(*, "classDef")= chr "SeuratObject:DimReduc"
#>  $ tsne:List of 9
#>   ..$ cell.embeddings           : num [1:80, 1:2] 0.868 -7.393 -28.206 16.348 1.911 ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : chr [1:80] "ATGCCAGAACGACT" "CATGGCCTGTGCAT" "GAACCTGATGAACC" "TGACTGGATTCTCA" ...
#>   .. .. ..$ : chr [1:2] "tSNE_1" "tSNE_2"
#>   ..$ feature.loadings          : num[0 , 0 ] 
#>   ..$ feature.loadings.projected: num[0 , 0 ] 
#>   ..$ assay.used                : chr "RNA"
#>   ..$ global                    : logi TRUE
#>   ..$ stdev                     : num(0) 
#>   ..$ jackstraw                 :List of 4
#>   .. ..$ empirical.p.values     : num[0 , 0 ] 
#>   .. ..$ fake.reduction.scores  : num[0 , 0 ] 
#>   .. ..$ empirical.p.values.full: num[0 , 0 ] 
#>   .. ..$ overall.p.values       : num[0 , 0 ] 
#>   .. ..- attr(*, "classDef")= chr "SeuratObject:JackStrawData"
#>   ..$ misc                      : list()
#>   ..$ key                       : chr "tSNE_"
#>   ..- attr(*, "classDef")= chr "SeuratObject:DimReduc"

IsS4List(pbmc.list)
#> [1] TRUE

pbmc2 <- ListToS4(pbmc.list)
pbmc2
#> An object of class Seurat 
#> 230 features across 80 samples within 1 assay 
#> Active assay: RNA (230 features, 20 variable features)
#>  3 layers present: counts, data, scale.data
#>  2 dimensional reductions calculated: pca, tsne
class(pbmc2)
#> [1] "Seurat"
#> attr(,"package")
#> [1] "SeuratObject"
Reductions(pbmc2)
#> [1] "pca"  "tsne"
validObject(pbmc2)
#> [1] TRUE