Get Feature Loadings

Pull feature loadings from a dimensional reduction

# S3 method for DimReduc
[(x, i, j, drop = FALSE, ...)

Arguments

x

A DimReduc object

i

Feature identifiers or indices

j

Dimension identifiers or indices

drop

Coerce the result to the lowest possible dimension; see drop for further details

...

Arguments passed to other methods

Value

Feature loadings for features i and dimensions j

Details

[ does not distinguish between projected and unprojected feature loadings; to select whether projected or unprojected loadings should be pulled, please use Loadings

See also

Examples

pca <- pbmc_small[["pca"]]
pca[1:10, 1:5]
#>               PC_1        PC_2          PC_3         PC_4         PC_5
#> PPBP    0.33832535  0.04095778  0.0292626090  0.031110335 -0.090420744
#> IGLL5  -0.03504289  0.05815335 -0.2990627165  0.547444540  0.214603428
#> VDAC3   0.11990482 -0.10994433 -0.0238602496  0.060151260 -0.809207588
#> CD1C   -0.04690284  0.19835522 -0.3509061724 -0.511121693 -0.130306281
#> AKR1C3 -0.03894635 -0.42880452  0.0884584725 -0.272743859  0.087791646
#> PF4     0.34392057  0.02474860 -0.0251951506 -0.012314114 -0.006725932
#> MYL9    0.29648344  0.05608286  0.0189977673 -0.034712318  0.262882991
#> GNLY   -0.06301277 -0.49660634  0.0327211725 -0.194851173 -0.013301495
#> TREML1  0.30589942  0.03516302 -0.0018399202 -0.001226459  0.108340799
#> CA2     0.31496381  0.04027236  0.0002813203 -0.017425402 -0.108099256