Matrix-like Subsetting for Logical Maps

Matrix-like Subsetting for Logical Maps

# S4 method for LogMap,missing,missing,ANY
[(x, i, j, ..., drop = FALSE)

# S4 method for LogMap,character,character,ANY
[(x, i, j, ..., drop = FALSE)

# S4 method for LogMap,character,missing,ANY
[(x, i, j, ..., drop = FALSE)

# S4 method for LogMap,missing,character,ANY
[(x, i, j, ..., drop = FALSE)

# S4 method for LogMap,numeric,missing,ANY
[(x, i, j, ..., drop = FALSE)

# S4 method for LogMap,missing,numeric,ANY
[(x, i, j, ..., drop = FALSE)

# S4 method for LogMap,numeric,numeric,ANY
[(x, i, j, ..., drop = FALSE)

Arguments

x

A LogMap object

i, j

Vectors of values (i) and observations (j) to pull from x

...

Arguments passed to other methods

drop

For matrices and arrays. If TRUE the result is coerced to the lowest possible dimension (see the examples). This only works for extracting elements, not for the replacement. See drop for further details.

Note

i is not reordable; passing a different order for i will return a subset with rows in the same order as x

Examples

map <- LogMap(letters[1:10])
map[['obs']] <- c(1, 3, 7)
map[['entry']] <- c(2, 7, 10)

map[]
#> A logical map for 10 values across 2 observations
map[1:5, 2L]
#> A logical map for 5 values across 1 observations
map[c("b", "c", "f"), "obs"]
#> A logical map for 3 values across 1 observations

# Pass `drop = TRUE` to cast to `matrix`
map[1:3, , drop = TRUE]
#>     obs entry
#> a  TRUE FALSE
#> b FALSE  TRUE
#> c  TRUE FALSE

# Note that `i` is non-reordable
rownames(map)[1:3]
#> [1] "a" "b" "c"
map[c("b", "c", "a"), , drop = TRUE]
#>     obs entry
#> a  TRUE FALSE
#> b FALSE  TRUE
#> c  TRUE FALSE