Add Subobjects

Add subobjects containing expression, dimensional reduction, or other containerized data to a Seurat object. Subobjects can be accessed with [[ and manipulated directly within the Seurat object or used independently

# S4 method for Seurat,character,missing,Assay
[[(x, i, j, ...) <- value

# S4 method for Seurat,character,missing,Assay5
[[(x, i, j, ...) <- value

# S4 method for Seurat,character,missing,DimReduc
[[(x, i, j, ...) <- value

# S4 method for Seurat,character,missing,Graph
[[(x, i, j, ...) <- value

# S4 method for Seurat,character,missing,Neighbor
[[(x, i, j, ...) <- value

# S4 method for Seurat,character,missing,SeuratCommand
[[(x, i, j, ...) <- value

# S4 method for Seurat,character,missing,SpatialImage
[[(x, i, j, ...) <- value

Arguments

x

A Seurat object

i

Name to add subobject as

j

Ignored

...

Ignored

value

A valid subobject (eg. a v3 or v5 assay, or a dimensional reduction)

Value

x with value added as i

See also