Seurat Methods

Methods for Seurat objects for generics defined in other packages

# S3 method for Seurat
.DollarNames(x, pattern = "")

# S3 method for Seurat
$(x, i, ...)

# S3 method for Seurat
$(x, i, ...) <- value

# S3 method for Seurat
[(x, i, j, ...)

# S3 method for Seurat
[[(x, i, ..., drop = FALSE)

# S3 method for Seurat
dim(x)

# S3 method for Seurat
dimnames(x)

# S3 method for Seurat
head(x, n = 10L, ...)

# S3 method for Seurat
merge(
  x = NULL,
  y = NULL,
  add.cell.ids = NULL,
  merge.data = TRUE,
  merge.dr = NULL,
  project = "SeuratProject",
  ...
)

# S3 method for Seurat
names(x)

# S3 method for Seurat
subset(
  x,
  subset,
  cells = NULL,
  features = NULL,
  idents = NULL,
  return.null = FALSE,
  ...
)

# S3 method for Seurat
tail(x, n = 10L, ...)

# S4 method for Seurat
[[(x, i, j, ...) <- value

# S4 method for Seurat
colMeans(x, na.rm = FALSE, dims = 1, ..., slot = "data")

# S4 method for Seurat
colSums(x, na.rm = FALSE, dims = 1, ..., slot = "data")

# S4 method for Seurat
rowMeans(x, na.rm = FALSE, dims = 1, ..., slot = "data")

# S4 method for Seurat
rowSums(x, na.rm = FALSE, dims = 1, ..., slot = "data")

# S4 method for Seurat
show(object)

Arguments

x, object
A Seurat object

pattern
A regular expression. Only matching names are returned.

i, features
Depends on the method

[, subset

Feature names or indices

$, $<-

Name of a single metadata column

[[, [[<-

Name of one or more metadata columns or an associated object; associated objects include Assay, DimReduc, Graph, SeuratCommand, or SpatialImage objects

...
Arguments passed to other methods

value
Additional metadata or associated objects to add; note: can pass NULL to remove metadata or an associated object

j, cells
Cell names or indices

drop
See drop

n
The number of rows of metadata to return

y
A single Seurat object or a list of Seurat objects

add.cell.ids
A character vector of length(x = c(x, y)); appends the corresponding values to the start of each objects' cell names

merge.data
Merge the data slots instead of just merging the counts (which requires renormalization); this is recommended if the same normalization approach was applied to all objects

merge.dr
Merge specified DimReducs that are present in all objects; will only merge the embeddings slots for the first N dimensions that are shared across all objects.

project
Project name for the Seurat object

subset
Logical expression indicating features/variables to keep

idents
A vector of identity classes to keep

return.null
If no cells are request, return a NULL; by default, throws an error

na.rm
logical. Should missing values (including NaN) be omitted from the calculations?

dims
completely ignored by the Matrix methods.

slot
Name of assay expression matrix to calculate column/row means/sums on

Value

$: metadata column i for object x; note: unlike [[, $ drops the shape of the metadata to return a vector instead of a data frame

$<-: object x with metadata value saved as i

[: object x with features i and cells j

[[: If i is missing, the metadata data frame; if i is a vector of metadata names, a data frame with the requested metadata, otherwise, the requested associated object

dim: The number of features (nrow) and cells (ncol) for the default assay; note: while the number of features changes depending on the active assay, the number of cells remains the same across all assays

dimnames: The feature (row) and cell (column) names; note: while the features change depending on the active assay, the cell names remain the same across all assays

head: The first n rows of cell-level metadata

merge: Merged object

names: The names of all Assay, DimReduc, Graph, and SpatialImage objects in the Seurat object

subset: A subsetted Seurat object

tail: The last n rows of cell-level metadata

[[<-: x with the metadata or associated objects added as i; if value is NULL, removes metadata or associated object i from object x

show: Prints summary to stdout and invisibly returns NULL

Functions

.DollarNames.Seurat: Autocompletion for $ access on a Seurat object

$.Seurat: Metadata access for Seurat objects

$<-.Seurat: Metadata setter for Seurat objects

[.Seurat: Simple subsetter for Seurat objects

[[.Seurat: Metadata and associated object accessor

dim.Seurat: Number of cells and features for the active assay

dimnames.Seurat: The cell and feature names for the active assay

head.Seurat: Get the first rows of cell-level metadata

merge.Seurat: Merge two or more Seurat objects together

names.Seurat: Common associated objects

subset.Seurat: Subset a Seurat object

tail.Seurat: Get the last rows of cell-level metadata

[[<-,Seurat-method: Add cell-level metadata or associated objects

colMeans,Seurat-method: Calculate colMeans on a Seurat object

colSums,Seurat-method: Calculate colSums on a Seurat object

rowMeans,Seurat-method: Calculate rowMeans on a rowMeans object

rowSums,Seurat-method: Calculate rowSums on a Seurat object

show,Seurat-method: Overview of a Seurat object

Merge Details

When merging Seurat objects, the merge procedure will merge the Assay level counts and potentially the data slots (depending on the merge.data parameter). It will also merge the cell-level meta data that was stored with each object and preserve the cell identities that were active in the objects pre-merge. The merge will optionally merge reductions depending on the values passed to merge.dr if they have the same name across objects. Here the embeddings slots will be merged and if there are differing numbers of dimensions across objects, only the first N shared dimensions will be merged. The feature loadings slots will be filled by the values present in the first object.The merge will not preserve graphs, logged commands, or feature-level metadata that were present in the original objects. If add.cell.ids isn't specified and any cell names are duplicated, cell names will be appended with _X, where X is the numeric index of the object in c(x, y).

See also

Examples

# Get metadata using `$'
head(pbmc_small$groups)

#> ATGCCAGAACGACT CATGGCCTGTGCAT GAACCTGATGAACC TGACTGGATTCTCA AGTCAGACTGCACA 
#>           "g2"           "g1"           "g2"           "g2"           "g2" 
#> TCTGATACACGTGT 
#>           "g1" 

# Add metadata using the `$' operator
set.seed(42)
pbmc_small$value <- sample(1:3, size = ncol(pbmc_small), replace = TRUE)
head(pbmc_small[["value"]])

#>                value
#> ATGCCAGAACGACT     1
#> CATGGCCTGTGCAT     1
#> GAACCTGATGAACC     1
#> TGACTGGATTCTCA     1
#> AGTCAGACTGCACA     2
#> TCTGATACACGTGT     2

# `[' examples
pbmc_small[VariableFeatures(object = pbmc_small), ]

#> An object of class Seurat 
#> 20 features across 80 samples within 1 assay 
#> Active assay: RNA (20 features, 20 variable features)
#>  2 dimensional reductions calculated: pca, tsne

pbmc_small[, 1:10]

#> An object of class Seurat 
#> 230 features across 10 samples within 1 assay 
#> Active assay: RNA (230 features, 20 variable features)
#>  2 dimensional reductions calculated: pca, tsne

# Get the cell-level metadata data frame
head(pbmc_small[[]])

#>                   orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.8
#> ATGCCAGAACGACT SeuratProject         70           47               0
#> CATGGCCTGTGCAT SeuratProject         85           52               0
#> GAACCTGATGAACC SeuratProject         87           50               1
#> TGACTGGATTCTCA SeuratProject        127           56               0
#> AGTCAGACTGCACA SeuratProject        173           53               0
#> TCTGATACACGTGT SeuratProject         70           48               0
#>                letter.idents groups RNA_snn_res.1 value
#> ATGCCAGAACGACT             A     g2             0     1
#> CATGGCCTGTGCAT             A     g1             0     1
#> GAACCTGATGAACC             B     g2             0     1
#> TGACTGGATTCTCA             A     g2             0     1
#> AGTCAGACTGCACA             A     g2             0     2
#> TCTGATACACGTGT             A     g1             0     2

# Pull specific metadata information
head(pbmc_small[[c("letter.idents", "groups")]])

#>                letter.idents groups
#> ATGCCAGAACGACT             A     g2
#> CATGGCCTGTGCAT             A     g1
#> GAACCTGATGAACC             B     g2
#> TGACTGGATTCTCA             A     g2
#> AGTCAGACTGCACA             A     g2
#> TCTGATACACGTGT             A     g1

head(pbmc_small[["groups", drop = TRUE]])

#> ATGCCAGAACGACT CATGGCCTGTGCAT GAACCTGATGAACC TGACTGGATTCTCA AGTCAGACTGCACA 
#>           "g2"           "g1"           "g2"           "g2"           "g2" 
#> TCTGATACACGTGT 
#>           "g1" 

# Get a sub-object (eg. an `Assay' or `DimReduc')
pbmc_small[["RNA"]]

#> Assay data with 230 features for 80 cells
#> Top 10 variable features:
#>  PPBP, IGLL5, VDAC3, CD1C, AKR1C3, PF4, MYL9, GNLY, TREML1, CA2 

pbmc_small[["pca"]]

#> A dimensional reduction object with key PC_ 
#>  Number of dimensions: 19 
#>  Projected dimensional reduction calculated:  TRUE 
#>  Jackstraw run: TRUE 
#>  Computed using assay: RNA 

# Get the number of features in an object
nrow(pbmc_small)

#> [1] 230

# Get the number of cells in an object
ncol(pbmc_small)

#> [1] 80

# Get the feature names of an object
rownames(pbmc_small)

#>   [1] "MS4A1"         "CD79B"         "CD79A"         "HLA-DRA"      
#>   [5] "TCL1A"         "HLA-DQB1"      "HVCN1"         "HLA-DMB"      
#>   [9] "LTB"           "LINC00926"     "FCER2"         "SP100"        
#>  [13] "NCF1"          "PPP3CC"        "EAF2"          "PPAPDC1B"     
#>  [17] "CD19"          "KIAA0125"      "CYB561A3"      "CD180"        
#>  [21] "RP11-693J15.5" "FAM96A"        "CXCR4"         "STX10"        
#>  [25] "SNHG7"         "NT5C"          "BANK1"         "IGLL5"        
#>  [29] "CD200"         "FCRLA"         "CD3D"          "NOSIP"        
#>  [33] "SAFB2"         "CD2"           "IL7R"          "PIK3IP1"      
#>  [37] "MPHOSPH6"      "KHDRBS1"       "MAL"           "CCR7"         
#>  [41] "THYN1"         "TAF7"          "LDHB"          "TMEM123"      
#>  [45] "CCDC104"       "EPC1"          "EIF4A2"        "CD3E"         
#>  [49] "TMUB1"         "BLOC1S4"       "ACSM3"         "TMEM204"      
#>  [53] "SRSF7"         "ACAP1"         "TNFAIP8"       "CD7"          
#>  [57] "TAGAP"         "DNAJB1"        "ASNSD1"        "S1PR4"        
#>  [61] "CTSW"          "GZMK"          "NKG7"          "IL32"         
#>  [65] "DNAJC2"        "LYAR"          "CST7"          "LCK"          
#>  [69] "CCL5"          "HNRNPH1"       "SSR2"          "DLGAP1-AS1"   
#>  [73] "GIMAP1"        "MMADHC"        "ZNF76"         "CD8A"         
#>  [77] "PTPN22"        "GYPC"          "HNRNPF"        "RPL7L1"       
#>  [81] "KLRG1"         "CRBN"          "SATB1"         "SIT1"         
#>  [85] "PMPCB"         "NRBP1"         "TCF7"          "HNRNPA3"      
#>  [89] "S100A8"        "S100A9"        "LYZ"           "CD14"         
#>  [93] "FCN1"          "TYROBP"        "ASGR1"         "NFKBIA"       
#>  [97] "TYMP"          "CTSS"          "TSPO"          "RBP7"         
#> [101] "CTSB"          "LGALS1"        "FPR1"          "VSTM1"        
#> [105] "BLVRA"         "MPEG1"         "BID"           "SMCO4"        
#> [109] "CFD"           "LINC00936"     "LGALS2"        "MS4A6A"       
#> [113] "FCGRT"         "LGALS3"        "NUP214"        "SCO2"         
#> [117] "IL17RA"        "IFI6"          "HLA-DPA1"      "FCER1A"       
#> [121] "CLEC10A"       "HLA-DMA"       "RGS1"          "HLA-DPB1"     
#> [125] "HLA-DQA1"      "RNF130"        "HLA-DRB5"      "HLA-DRB1"     
#> [129] "CST3"          "IL1B"          "POP7"          "HLA-DQA2"     
#> [133] "CD1C"          "GSTP1"         "EIF3G"         "VPS28"        
#> [137] "LY86"          "ZFP36L1"       "ZNF330"        "ANXA2"        
#> [141] "GRN"           "CFP"           "HSP90AA1"      "FUOM"         
#> [145] "LST1"          "AIF1"          "PSAP"          "YWHAB"        
#> [149] "MYO1G"         "SAT1"          "RGS2"          "SERPINA1"     
#> [153] "IFITM3"        "FCGR3A"        "LILRA3"        "S100A11"      
#> [157] "FCER1G"        "TNFRSF1B"      "IFITM2"        "WARS"         
#> [161] "IFI30"         "MS4A7"         "C5AR1"         "HCK"          
#> [165] "COTL1"         "LGALS9"        "CD68"          "RP11-290F20.3"
#> [169] "RHOC"          "CARD16"        "LRRC25"        "COPS6"        
#> [173] "ADAR"          "PPBP"          "GPX1"          "TPM4"         
#> [177] "PF4"           "SDPR"          "NRGN"          "SPARC"        
#> [181] "GNG11"         "CLU"           "HIST1H2AC"     "NCOA4"        
#> [185] "GP9"           "FERMT3"        "ODC1"          "CD9"          
#> [189] "RUFY1"         "TUBB1"         "TALDO1"        "TREML1"       
#> [193] "NGFRAP1"       "PGRMC1"        "CA2"           "ITGA2B"       
#> [197] "MYL9"          "TMEM40"        "PARVB"         "PTCRA"        
#> [201] "ACRBP"         "TSC22D1"       "VDAC3"         "GZMB"         
#> [205] "GZMA"          "GNLY"          "FGFBP2"        "AKR1C3"       
#> [209] "CCL4"          "PRF1"          "GZMH"          "XBP1"         
#> [213] "GZMM"          "PTGDR"         "IGFBP7"        "TTC38"        
#> [217] "KLRD1"         "ARHGDIA"       "IL2RB"         "CLIC3"        
#> [221] "PPP1R18"       "CD247"         "ALOX5AP"       "XCL2"         
#> [225] "C12orf75"      "RARRES3"       "PCMT1"         "LAMP1"        
#> [229] "SPON2"         "S100B"        

# Get the cell names of an object
colnames(pbmc_small)

#>  [1] "ATGCCAGAACGACT" "CATGGCCTGTGCAT" "GAACCTGATGAACC" "TGACTGGATTCTCA"
#>  [5] "AGTCAGACTGCACA" "TCTGATACACGTGT" "TGGTATCTAAACAG" "GCAGCTCTGTTTCT"
#>  [9] "GATATAACACGCAT" "AATGTTGACAGTCA" "AGGTCATGAGTGTC" "AGAGATGATCTCGC"
#> [13] "GGGTAACTCTAGTG" "CATGAGACACGGGA" "TACGCCACTCCGAA" "CTAAACCTGTGCAT"
#> [17] "GTAAGCACTCATTC" "TTGGTACTGAATCC" "CATCATACGGAGCA" "TACATCACGCTAAC"
#> [21] "TTACCATGAATCGC" "ATAGGAGAAACAGA" "GCGCACGACTTTAC" "ACTCGCACGAAAGT"
#> [25] "ATTACCTGCCTTAT" "CCCAACTGCAATCG" "AAATTCGAATCACG" "CCATCCGATTCGCC"
#> [29] "TCCACTCTGAGCTT" "CATCAGGATGCACA" "CTAAACCTCTGACA" "GATAGAGAAGGGTG"
#> [33] "CTAACGGAACCGAT" "AGATATACCCGTAA" "TACTCTGAATCGAC" "GCGCATCTTGCTCC"
#> [37] "GTTGACGATATCGG" "ACAGGTACTGGTGT" "GGCATATGCTTATC" "CATTACACCAACTG"
#> [41] "TAGGGACTGAACTC" "GCTCCATGAGAAGT" "TACAATGATGCTAG" "CTTCATGACCGAAT"
#> [45] "CTGCCAACAGGAGC" "TTGCATTGAGCTAC" "AAGCAAGAGCTTAG" "CGGCACGAACTCAG"
#> [49] "GGTGGAGATTACTC" "GGCCGATGTACTCT" "CGTAGCCTGTATGC" "TGAGCTGAATGCTG"
#> [53] "CCTATAACGAGACG" "ATAAGTTGGTACGT" "AAGCGACTTTGACG" "ACCAGTGAATACCG"
#> [57] "ATTGCACTTGCTTT" "CTAGGTGATGGTTG" "GCACTAGACCTTTA" "CATGCGCTAGTCAC"
#> [61] "TTGAGGACTACGCA" "ATACCACTCTAAGC" "CATATAGACTAAGC" "TTTAGCTGTACTCT"
#> [65] "GACATTCTCCACCT" "ACGTGATGCCATGA" "ATTGTAGATTCCCG" "GATAGAGATCACGA"
#> [69] "AATGCGTGGACGGA" "GCGTAAACACGGTT" "ATTCAGCTCATTGG" "GGCATATGGGGAGT"
#> [73] "ATCATCTGACACCA" "GTCATACTTCGCCT" "TTACGTACGTTCAG" "GAGTTGTGGTAGCT"
#> [77] "GACGCTCTCTCTCG" "AGTCTTACTTCGGA" "GGAACACTTCAGAC" "CTTGATTGATCTTC"

# Get the first 10 rows of cell-level metadata
head(pbmc_small)

#>                   orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.8
#> ATGCCAGAACGACT SeuratProject         70           47               0
#> CATGGCCTGTGCAT SeuratProject         85           52               0
#> GAACCTGATGAACC SeuratProject         87           50               1
#> TGACTGGATTCTCA SeuratProject        127           56               0
#> AGTCAGACTGCACA SeuratProject        173           53               0
#> TCTGATACACGTGT SeuratProject         70           48               0
#> TGGTATCTAAACAG SeuratProject         64           36               0
#> GCAGCTCTGTTTCT SeuratProject         72           45               0
#> GATATAACACGCAT SeuratProject         52           36               0
#> AATGTTGACAGTCA SeuratProject        100           41               0
#>                letter.idents groups RNA_snn_res.1 value
#> ATGCCAGAACGACT             A     g2             0     1
#> CATGGCCTGTGCAT             A     g1             0     1
#> GAACCTGATGAACC             B     g2             0     1
#> TGACTGGATTCTCA             A     g2             0     1
#> AGTCAGACTGCACA             A     g2             0     2
#> TCTGATACACGTGT             A     g1             0     2
#> TGGTATCTAAACAG             A     g1             0     2
#> GCAGCTCTGTTTCT             A     g1             0     1
#> GATATAACACGCAT             A     g1             0     3
#> AATGTTGACAGTCA             A     g1             0     3

# `merge' examples
# merge two objects
merge(pbmc_small, y = pbmc_small)

#> Warning: Some cell names are duplicated across objects provided. Renaming to enforce unique cell names.

#> An object of class Seurat 
#> 230 features across 160 samples within 1 assay 
#> Active assay: RNA (230 features, 0 variable features)

# to merge more than two objects, pass one to x and a list of objects to y
merge(pbmc_small, y = c(pbmc_small, pbmc_small))

#> Warning: Some cell names are duplicated across objects provided. Renaming to enforce unique cell names.

#> An object of class Seurat 
#> 230 features across 240 samples within 1 assay 
#> Active assay: RNA (230 features, 0 variable features)

names(pbmc_small)

#> [1] "RNA"     "RNA_snn" "pca"     "tsne"   

# `subset' examples
subset(pbmc_small, subset = MS4A1 > 4)

#> An object of class Seurat 
#> 230 features across 10 samples within 1 assay 
#> Active assay: RNA (230 features, 20 variable features)
#>  2 dimensional reductions calculated: pca, tsne

subset(pbmc_small, subset = `DLGAP1-AS1` > 2)

#> An object of class Seurat 
#> 230 features across 4 samples within 1 assay 
#> Active assay: RNA (230 features, 20 variable features)
#>  2 dimensional reductions calculated: pca, tsne

subset(pbmc_small, idents = '0', invert = TRUE)

#> An object of class Seurat 
#> 230 features across 44 samples within 1 assay 
#> Active assay: RNA (230 features, 20 variable features)
#>  2 dimensional reductions calculated: pca, tsne

subset(pbmc_small, subset = MS4A1 > 3, slot = 'counts')

#> An object of class Seurat 
#> 230 features across 3 samples within 1 assay 
#> Active assay: RNA (230 features, 20 variable features)
#>  2 dimensional reductions calculated: pca, tsne

subset(pbmc_small, features = VariableFeatures(object = pbmc_small))

#> An object of class Seurat 
#> 20 features across 80 samples within 1 assay 
#> Active assay: RNA (20 features, 20 variable features)
#>  2 dimensional reductions calculated: pca, tsne

# Get the last 10 rows of cell-level metadata
tail(pbmc_small)

#>                   orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.8
#> ATTCAGCTCATTGG SeuratProject        212           38               0
#> GGCATATGGGGAGT SeuratProject        172           29               0
#> ATCATCTGACACCA SeuratProject        168           37               0
#> GTCATACTTCGCCT SeuratProject        210           33               0
#> TTACGTACGTTCAG SeuratProject        228           39               0
#> GAGTTGTGGTAGCT SeuratProject        527           47               0
#> GACGCTCTCTCTCG SeuratProject        202           30               0
#> AGTCTTACTTCGGA SeuratProject        157           29               0
#> GGAACACTTCAGAC SeuratProject        150           30               0
#> CTTGATTGATCTTC SeuratProject        233           76               1
#>                letter.idents groups RNA_snn_res.1 value
#> ATTCAGCTCATTGG             A     g2             0     2
#> GGCATATGGGGAGT             A     g1             0     1
#> ATCATCTGACACCA             A     g2             0     2
#> GTCATACTTCGCCT             A     g2             0     2
#> TTACGTACGTTCAG             A     g1             0     2
#> GAGTTGTGGTAGCT             A     g1             0     1
#> GACGCTCTCTCTCG             A     g2             0     2
#> AGTCTTACTTCGGA             A     g1             0     1
#> GGAACACTTCAGAC             A     g2             0     2
#> CTTGATTGATCTTC             B     g1             1     3

head(colMeans(pbmc_small))

#> ATGCCAGAACGACT CATGGCCTGTGCAT GAACCTGATGAACC TGACTGGATTCTCA AGTCAGACTGCACA 
#>       1.070119       1.156175       1.102886       1.200849       1.051583 
#> TCTGATACACGTGT 
#>       1.091485 

head(colSums(pbmc_small))

#> ATGCCAGAACGACT CATGGCCTGTGCAT GAACCTGATGAACC TGACTGGATTCTCA AGTCAGACTGCACA 
#>       246.1273       265.9203       253.6638       276.1952       241.8640 
#> TCTGATACACGTGT 
#>       251.0415 

head(rowMeans(pbmc_small))

#>     MS4A1     CD79B     CD79A   HLA-DRA     TCL1A  HLA-DQB1 
#> 0.7890259 1.3545382 0.8719449 3.9500607 0.5661287 1.7583174 

head(rowSums(pbmc_small))

#>     MS4A1     CD79B     CD79A   HLA-DRA     TCL1A  HLA-DQB1 
#>  63.12207 108.36306  69.75559 316.00486  45.29029 140.66539