Merge Seurat Objects

Merge two or more objects.

# S3 method for Seurat
merge(
  x = NULL,
  y = NULL,
  add.cell.ids = NULL,
  merge.data = TRUE,
  merge.dr = NULL,
  project = "SeuratProject",
  ...
)

Arguments

x
Object

y
Object (or a list of multiple objects)

add.cell.ids
A character vector of length(x = c(x, y)). Appends the corresponding values to the start of each objects' cell names.

merge.data
Merge the data slots instead of just merging the counts (which requires renormalization). This is recommended if the same normalization approach was applied to all objects.

merge.dr
Merge specified DimReducs that are present in all objects. Will only merge the embeddings slots for the first N dimensions that are shared across all objects.

project
Project name for the Seurat object

...
Arguments passed to other methods

Value

Merged object

Details

When merging Seurat objects, the merge procedure will merge the Assay level counts and potentially the data slots (depending on the merge.data parameter). It will also merge the cell-level meta data that was stored with each object and preserve the cell identities that were active in the objects pre-merge. The merge will optionally merge reductions depending on the values passed to merge.dr if they have the same name across objects. Here the embeddings slots will be merged and if there are differing numbers of dimensions across objects, only the first N shared dimensions will be merged. The feature loadings slots will be filled by the values present in the first object.The merge will not preserve graphs, logged commands, or feature-level metadata that were present in the original objects. If add.cell.ids isn't specified and any cell names are duplicated, cell names will be appended with _X, where X is the numeric index of the object in c(x, y).

Examples

# merge two objects
merge(x = pbmc_small, y = pbmc_small)

#> Error in merge(x = pbmc_small, y = pbmc_small): object 'pbmc_small' not found

# to merge more than two objects, pass one to x and a list of objects to y
merge(x = pbmc_small, y = c(pbmc_small, pbmc_small))

#> Error in merge(x = pbmc_small, y = c(pbmc_small, pbmc_small)): object 'pbmc_small' not found